New modular menu for options. See OKMenu struct and it's elements.
Create char arrays for menu text and parameters, and define pointers and panel counts. for each panel, define Option counts and the text of each option and the parameters as well. Options menu will allow users to iterate through each panel, displaying the options and parameters for each. When toggling an option, it will flip the accompanying char in the defined Parameter array.

Need to fix the overflow; currently printing past 4 options.